Move a Brig fed API notif endpoint to a module
