+++ libs/wire-api-federation/wire-api-federation.cabal
@@ -18,6 +18,7 @@ library
   exposed-modules:
     Wire.API.Federation.API
     Wire.API.Federation.API.Brig
+    Wire.API.Federation.API.Brig.Notifications
     Wire.API.Federation.API.Cargohold
     Wire.API.Federation.API.Common
     Wire.API.Federation.API.Galley
